Refine snmalloc integration and PALs
This commit includes further refinements to the snmalloc integration:

- Verified and corrected the type of `gMappedPagesCount` in the kernel PAL
  (`pal_haiku_kernel.h`) to `int32`.
- Confirmed that explicit compilation of core `snmalloc` .cc files is likely
  unnecessary for libroot and kernel, as the Haiku-specific API/hook files
  will include `snmalloc` headers, leveraging its header-heavy/template-based design.
- Made diagnostic `Echo` messages in relevant Jamfiles more concise.

These changes build upon the initial integration of `snmalloc` as a selectable
allocator for both libroot and the kernel.

+++ src/system/kernel/alloc/snmalloc/pal_haiku_kernel.h
@@ -28,7 +28,7 @@
 #include <slab/Slab.h> // For object_cache_alloc (used by old PAL version, now only for HaikuKernelSubMapping if not using internal pool)
 #include <kernel/thread.h> // For spinlock, disable_interrupts, etc.
 
-extern int64 gMappedPagesCount; // For tracking mapped pages globally in kernel
+extern int32 gMappedPagesCount; // For tracking mapped pages globally in kernel
 
 
 // Forward declaration
